change float32 to float64, replace validateHelper

+++ infrastructure/handlers/interface.go
@@ -1,8 +1,8 @@
 package handlers
 
 type TransactionService interface {
-	AddToBalance(userId int, amount float32) (float32, error)
-	AddTransfer(fromId int, toId int, amount float32) (float32, float32, error)
+	AddToBalance(userId int, amount float64) (float64, error)
+	AddTransfer(fromId int, toId int, amount float64) (float64, float64, error)
 }
 
 type UserCreator interface {
