+++ infrastructure/handlers/type.go
@@ -2,11 +2,11 @@ package handlers
 
 type AddToBalanceRequest struct {
 	UserId int     `json:"userId"`
-	Amount float32 `json:"amount"`
+	Amount float64 `json:"amount"`
 }
 
 type AddTransferRequest struct {
 	FromId int     `json:"fromId"`
 	ToId   int     `json:"toId"`
-	Amount float32 `json:"amount"`
+	Amount float64 `json:"amount"`
 }
