+++ infrastructure/storage/type.go
@@ -2,5 +2,5 @@ package storage
 
 type User struct {
 	Id      int
-	Balance float32
+	Balance float64
 }
